+++ ContentPatcher/docs/release-notes.md
@@ -12,6 +12,7 @@ When releasing a format change, don't forget to update the smapi.io/json schema!
 ## Upcoming release
 * Added [`PatchMode: Mask`](author-guide/action-editimage.md) to apply transparency masks to images (thanks to 
 PinkSerenity!).
+* Fixed error when a content pack uses `Fields` to edit a field within a null parent field. Content Patcher will now try to initialize the entry if possible, else log a warning.
 
 ## 2.8.1
 Released 26 October 2025 for SMAPI 4.1.10 or later.